Merge branch 'fix/thread-form' into 'develop'

See merge request yeneua.dev/final-pjt!27

+++ backend/books/views.py
@@ -3,6 +3,7 @@ from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 from rest_framework import status
 from .models import Book, Category, Thread, Comment
+from .permissions import IsAuthenticatedOrReadOnly
 from .serializers import (
     BookSerializer,
     CategorySerializer,
@@ -51,8 +52,8 @@ def book_detail(request, book_id):
         serializer = BookSerializer(book)
         return Response(serializer.data)
 
-@api_view(["GET", "DELETE", "PUT",])
-# @permission_classes([IsAuthenticated])
+@api_view(["GET", "DELETE", "PUT"])
+@permission_classes([IsAuthenticatedOrReadOnly])
 def thread_detail(request, thread_id):
     thread = Thread.objects.get(id=thread_id)
 
@@ -60,14 +61,17 @@ def thread_detail(request, thread_id):
         serializer = ThreadDetailSerializer(thread)
         return Response(serializer.data)
     elif request.method == "DELETE":
+        if thread.user != request.user:
+            return Response({"detail": "권한이 없습니다."}, status=status.HTTP_403_FORBIDDEN)
         thread.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
     elif request.method == "PUT":
-        serializer = ThreadSerializer(thread, data=request.data, partial=True)
+        if thread.user != request.user:
+            return Response({"detail": "권한이 없습니다."}, status=status.HTTP_403_FORBIDDEN)
+        serializer = ThreadCreateSerializer(thread, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
-
 
 @api_view(["POST"])
 @permission_classes([IsAuthenticated])
@@ -79,7 +83,6 @@ def thread_likes(request, thread_id):
         thread.likes.add(request.user)
     serializer = ThreadSerializer(thread, context={"request": request})
     return Response(serializer.data)
-
 
 @api_view(["GET", "POST"])
 def comment_list(request, thread_id):
